Update delete user router - check if order exist logic added

+++ orders/src/routes/orders/delete.ts
@@ -5,6 +5,7 @@ import {
     STATUS_NO_CONTENT,
 } from "../../constants/data";
 import {OrderService} from "../../services/orderService";
+import {NotFoundRequestError} from "@dmstickets/common";
 
 const service = new OrderService()
 
@@ -12,6 +13,10 @@ const router = express.Router()
 router.delete("/api/orders/:id", [
 ], async (req: Request, res: Response) => {
     const { id } = req.params
+    const order = await service
+        .setId(id)
+        .getOrder()
+    if (!order) throw new NotFoundRequestError("Order didn't exists!")
     await service
         .setId(id)
         .deleteOrder()
